Populate postgres with switch parameter
When ingesting the tree of triples, N tokens up to switch_parameter are
ingested with the key,children paradigm while tokens at levels higher
than the switch_parameter are ingested as a serialized subtree that will
be loaded entirely in memory during generation.

This is to reduce the number of rows in the db when the subtree size is
easy to manage.

The estimate script objective is to help estimate a good switch_parameter
value.