Improve BI dataset
Translate the relations in a more natural form. Add inverse relations
and reflexive ones.

Modify BI prompt so that it assumes that no fact means the fact is
false and to avoid using descriptions (not present in this KB).

Move prompt template in dataset config to easier modify prompts for
different datasets.